+++ src/types/index.ts
@@ -92,6 +92,7 @@ export type {
   CDeposit,
   CWithdraw,
   CoreEvmTransfer,
+  SendToEvmWithData,
 } from './account.js';
 export type { TokenDelegate, ClaimRewards, LinkStakingUser } from './staking.js';
 export type { TwapOrder, TwapCancel } from './twap.js';